Fix ompi-server operations
(cherry picked from commit 8c2a06477c85ae617c256f71ae40c9f7340a7ee6)
