+++ orte/orted/help-orted.txt
@@ -80,3 +80,10 @@ This is usually caused by a large job that encounters significant
 delays across the cluster when starting the application processes.
 Your job may terminate as a result of this problem. You may want to
 adjust the MCA parameter pmix_server_max_reqs and try again.
+#
+[noserver]
+A publish/lookup server was provided, but we were unable to connect
+to it - please check the connection info and ensure the server
+is alive:
+
+  Connection:  %s
