+++ orte/orted/pmix/pmix_server.c
@@ -296,94 +296,6 @@ int pmix_server_init(void)
     }
     OPAL_LIST_DESTRUCT(&info);
 
-    /* if the universal server wasn't specified, then we use
-     * our own HNP for that purpose */
-    if (NULL == orte_pmix_server_globals.server_uri) {
-        orte_pmix_server_globals.server = *ORTE_PROC_MY_HNP;
-    } else {
-        char *server;
-        opal_buffer_t buf;
-        if (0 == strncmp(orte_pmix_server_globals.server_uri, "file", strlen("file")) ||
-            0 == strncmp(orte_pmix_server_globals.server_uri, "FILE", strlen("FILE"))) {
-            char input[1024], *filename;
-            FILE *fp;
-
-            /* it is a file - get the filename */
-            filename = strchr(orte_pmix_server_globals.server_uri, ':');
-            if (NULL == filename) {
-                /* filename is not correctly formatted */
-                orte_show_help("help-orterun.txt", "orterun:ompi-server-filename-bad", true,
-                               orte_basename, orte_pmix_server_globals.server_uri);
-                return ORTE_ERR_BAD_PARAM;
-            }
-            ++filename; /* space past the : */
-
-            if (0 >= strlen(filename)) {
-                /* they forgot to give us the name! */
-                orte_show_help("help-orterun.txt", "orterun:ompi-server-filename-missing", true,
-                               orte_basename, orte_pmix_server_globals.server_uri);
-                return ORTE_ERR_BAD_PARAM;
-            }
-
-            /* open the file and extract the uri */
-            fp = fopen(filename, "r");
-            if (NULL == fp) { /* can't find or read file! */
-                orte_show_help("help-orterun.txt", "orterun:ompi-server-filename-access", true,
-                               orte_basename, orte_pmix_server_globals.server_uri);
-                return ORTE_ERR_BAD_PARAM;
-            }
-            if (NULL == fgets(input, 1024, fp)) {
-                /* something malformed about file */
-                fclose(fp);
-                orte_show_help("help-orterun.txt", "orterun:ompi-server-file-bad", true,
-                               orte_basename, orte_pmix_server_globals.server_uri,
-                               orte_basename);
-                return ORTE_ERR_BAD_PARAM;
-            }
-            fclose(fp);
-            input[strlen(input)-1] = '\0';  /* remove newline */
-            server = strdup(input);
-        } else {
-            server = strdup(orte_pmix_server_globals.server_uri);
-        }
-        /* setup our route to the server */
-        OBJ_CONSTRUCT(&buf, opal_buffer_t);
-        opal_dss.pack(&buf, &server, 1, OPAL_STRING);
-        if (ORTE_SUCCESS != (rc = orte_rml_base_update_contact_info(&buf))) {
-            ORTE_ERROR_LOG(rc);
-            ORTE_UPDATE_EXIT_STATUS(ORTE_ERROR_DEFAULT_EXIT_CODE);
-            return rc;
-        }
-        OBJ_DESTRUCT(&buf);
-        /* parse the URI to get the server's name */
-        if (ORTE_SUCCESS != (rc = orte_rml_base_parse_uris(server, &orte_pmix_server_globals.server, NULL))) {
-            ORTE_ERROR_LOG(rc);
-            return rc;
-        }
-        /* check if we are to wait for the server to start - resolves
-         * a race condition that can occur when the server is run
-         * as a background job - e.g., in scripts
-         */
-        if (orte_pmix_server_globals.wait_for_server) {
-            /* ping the server */
-            struct timeval timeout;
-            timeout.tv_sec = orte_pmix_server_globals.timeout;
-            timeout.tv_usec = 0;
-            if (ORTE_SUCCESS != (rc = orte_rml.ping(orte_mgmt_conduit, server, &timeout))) {
-                /* try it one more time */
-                if (ORTE_SUCCESS != (rc = orte_rml.ping(orte_mgmt_conduit, server, &timeout))) {
-                    /* okay give up */
-                    orte_show_help("help-orterun.txt", "orterun:server-not-found", true,
-                                   orte_basename, server,
-                                   (long)orte_pmix_server_globals.timeout,
-                                   ORTE_ERROR_NAME(rc));
-                    ORTE_UPDATE_EXIT_STATUS(ORTE_ERROR_DEFAULT_EXIT_CODE);
-                    return rc;
-                }
-            }
-        }
-    }
-
     return rc;
 }
 
@@ -716,8 +628,9 @@ OBJ_CLASS_INSTANCE(orte_pmix_server_op_caddy_t,
 static void rqcon(pmix_server_req_t *p)
 {
     p->operation = NULL;
-    p->target = *ORTE_NAME_INVALID;
+    p->range = OPAL_PMIX_RANGE_SESSION;
     p->proxy = *ORTE_NAME_INVALID;
+    p->target = *ORTE_NAME_INVALID;
     p->timeout = orte_pmix_server_globals.timeout;
     p->jdata = NULL;
     OBJ_CONSTRUCT(&p->msg, opal_buffer_t);
